third_party/WebKit/LayoutTests/netinfo/basic-operation-expected.txt

Tests the basic operation of NetInfo.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.connection is defined.
PASS navigator.connection.type is defined.
PASS typeof connection.type is "string"
PASS connection.type is newConnectionType
PASS successfullyParsed is true

TEST COMPLETE
